Move the ancestry verification into the lib.rs file
