+++ srml/bridge/src/lib.rs
@@ -130,6 +130,7 @@ decl_error! {
 		InvalidStorageProof,
 		InvalidValidatorSetProof,
 		ValidatorSetMismatch,
+		AncestorNotFound,
 	}
 }
 
@@ -158,6 +159,39 @@ impl<T: Trait> Module<T> {
 			Err(Error::ValidatorSetMismatch)
 		}
 	}
+}
+
+// A naive way to check whether a `child` header is an ancestor
+// of an `ancestor` header. For this it requires a proof which
+// is a header chain, This could be updated to use something like
+// Log2 Ancestors (#2053) in the future.
+fn verify_ancestry<H>(proof: Vec<H>, ancestor: H, child: H) -> std::result::Result<(), Error>
+where
+	H: Header
+{
+	let mut curr_header = &proof[0];
+	if curr_header.hash() != child.hash() {
+		return Err(Error::AncestorNotFound);
+	}
+
+	let mut parent_hash = curr_header.parent_hash();
+
+	// If we find that the header's parent hash matches our ancestor's hash we're done
+	for i in 1..proof.len() {
+		curr_header = &proof[i];
+
+		// Need to check that blocks are actually related
+		if curr_header.hash() != *parent_hash {
+			break;
+		}
+
+		parent_hash = curr_header.parent_hash();
+		if *parent_hash == ancestor.hash() {
+			return Ok(())
+		}
+	}
+
+	Err(Error::AncestorNotFound)
 }
 
 #[cfg(test)]
@@ -302,5 +336,91 @@ mod tests {
 
 			assert_eq!(MockBridge::num_bridges(), 1);
 		});
+	}
+
+	fn get_related_block_headers() -> (Header, Header, Header) {
+		let grandparent = Header {
+			parent_hash: H256::default(),
+			number: 1,
+			state_root: H256::default(),
+			extrinsics_root: H256::default(),
+			digest: Digest::default(),
+		};
+
+		let parent = Header {
+			parent_hash: grandparent.hash(),
+			number: grandparent.number() + 1,
+			state_root: H256::default(),
+			extrinsics_root: H256::default(),
+			digest: Digest::default(),
+		};
+
+		let child = Header {
+			parent_hash: parent.hash(),
+			number: parent.number() + 1,
+			state_root: H256::default(),
+			extrinsics_root: H256::default(),
+			digest: Digest::default(),
+		};
+
+		(grandparent, parent, child)
+	}
+
+	#[test]
+	fn check_that_child_is_ancestor_of_grandparent() {
+		let (grandparent, parent, child) = get_related_block_headers();
+
+		let mut proof = Vec::new();
+		proof.push(child.clone());
+		proof.push(parent);
+		proof.push(grandparent.clone());
+
+		assert_ok!(verify_ancestry(proof, grandparent, child));
+	}
+
+	#[test]
+	fn check_that_child_ancestor_is_not_correct() {
+		let (grandparent, parent, child) = get_related_block_headers();
+
+		let mut proof = Vec::new();
+		proof.push(child.clone());
+		proof.push(parent);
+		proof.push(grandparent.clone());
+
+		let fake_grandparent = Header {
+			parent_hash: H256::from_slice(&[1u8; 32]),
+			number: 42,
+			state_root: H256::default(),
+			extrinsics_root: H256::default(),
+			digest: Digest::default(),
+		};
+
+		assert_err!(
+			verify_ancestry(proof, fake_grandparent, child),
+			Error::AncestorNotFound
+		);
+	}
+
+	#[test]
+	fn checker_fails_if_given_invalid_proof() {
+		let (grandparent, parent, child) = get_related_block_headers();
+		let fake_ancestor = Header {
+			parent_hash: H256::from_slice(&[1u8; 32]),
+			number: 42,
+			state_root: H256::default(),
+			extrinsics_root: H256::default(),
+			digest: Digest::default(),
+		};
+
+		let mut invalid_proof = Vec::new();
+		invalid_proof.push(child.clone());
+		invalid_proof.push(fake_ancestor);
+		invalid_proof.push(parent);
+		invalid_proof.push(grandparent.clone());
+
+		assert_err!(
+			verify_ancestry(invalid_proof, grandparent, child),
+			Error::AncestorNotFound
+		);
 	}
 }
